Fix a false positive caused by Core Data deallocating NSManagedObjects on a background queue.
A disassembly of -[NSManagedObjectContext dealloc] indicates that if the context has more than 300 objects that it has references to (and it doesn't have a private dispatch queue already), then it will deallocate the objects on a new private queue.  This would normally be considered unsafe access, but it's Core Data doing this, so we must assume it to be safe.

The false positive has been fixed by not validating concurrency for -dealloc method calls.  This shouldn't miss any unsafe concurrency because -release would normally immediately precede a -dealloc, and we validate calls to -release.