Fix libdispatch function interposing on iOS.
The problem was that on iOS `dlsym(RTLD_DEFAULT, "dispatch_sync")` wasn't returning the address for the function dispatch_sync, but instead a stub in libdispatch.  This stub looks up the function address in the lazy function table and then jumps to it.  The problem is that fishhook was modifying the lazy function table to point back at our version.  So we'd get infinite recursion with our wrapped dispatch_sync calling itself.

The fix is to get fishhook to *not* rebind the symbols in libdispatch.  That way, when we call the 'original' dispatch_sync function, the stub (which is in lib dispatch) will jump to the real dispatch_sync.

This commit also interposes the block-based dispatch_sync and dispatch_barrier_sync calls.  They aren't used by CoreData, but we should do it just in case.