Add support for using a serial NSOperationQueue as your confinement mechanism.
GDCoreDataConcurrencyDebugging works by verifying the API contract you sign up to by the concurrency type you specify to `-[NSManagedObjectContext initWithConcurrencyType:]`.  In theory, it will probably be valid to use an `NSConfinementConcurrencyType` context on a serial NSOperationQueue (or a serial dispatch queue), although CoreData makes no guarantees that this will be safe.  With this commit, you can specify `GDOperationQueueConcurrencyType` as the argument to `-[NSManagedObjectContext initWithConcurrencyType:]` to make GDCoreDataConcurrencyDebugging verify this use.  You will need to `#import <GDCoreDataConcurrencyDebugging/GDConcurrencyCheckingManagedObject.h>` to use this symbol.

Fixes #11.